Isolate fetchHardforkVersion from ProviderService
